requirements.txt
# Configuration
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
psutil>=5.9.0

# Trading
alpaca-trade-api>=3.1.1
python-dotenv>=1.0.0
websockets>=10.0,<11.0

# Database
psycopg2-binary>=2.9.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0

# Dashboard
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3

# Data Collection
yfinance>=0.2.18,<0.2.65 

# Workflow Orchestration
prefect>=3.4.14